chore: abstract utils to their own files
