+++ .changeset/fifty-tables-yell.md
@@ -1,0 +1,5 @@
+---
+"vtta": patch
+---
+
+Abstract utils out into their own files
